refactor(devtools): move ContainerType, DebugSignalGraphNode, Descriptor, and PropType to primitives
`DebugSignalGraphNode` is the return type of the `ng` global `getSignalGraph` function. It will be used by Client-Only Wiz and other frameworks that implement `getSignalGraph`. Movig it to primitives allows those frameworks to import and reference the type.
