+++ devtools/projects/ng-devtools-backend/src/lib/state-serializer/BUILD.bazel
@@ -22,6 +22,7 @@ ts_test_library(
     ],
     deps = [
         ":state-serializer",
+        "//:node_modules/@angular/core",
         "//devtools/projects/protocol",
     ],
 )
