+++ devtools/projects/ng-devtools-backend/src/lib/state-serializer/prop-type.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {PropType} from '../../../../protocol';
+import {ɵPropType as PropType} from '@angular/core';
 
 import {isSignal, safelyReadSignalValue} from '../utils/general';
