+++ devtools/projects/ng-devtools-backend/src/lib/state-serializer/serialized-descriptor-factory.ts
@@ -6,7 +6,12 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {ContainerType, Descriptor, NestedProp, PropType} from '../../../../protocol';
+import {
+  type ɵContainerType as ContainerType,
+  type ɵDescriptor as Descriptor,
+  ɵPropType as PropType,
+} from '@angular/core';
+import {NestedProp} from '../../../../protocol';
 
 import {isSignal, safelyReadSignalValue, unwrapSignal} from '../utils/general';
