+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/property-resolver/directive-property-resolver.ts
@@ -7,8 +7,8 @@
  */
 
 import {FlatTreeControl} from '@angular/cdk/tree';
+import type {ɵDescriptor as Descriptor} from '@angular/core';
 import {
-  Descriptor,
   DirectiveMetadata,
   DirectivePosition,
   Events,
