+++ devtools/projects/ng-devtools/src/lib/shared/object-tree-explorer/object-tree-types.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {Descriptor} from '../../../../../protocol';
+import type {ɵDescriptor as Descriptor} from '@angular/core';
 
 /**
  * An object-tree-exlorer data source unit that represents an object.
