+++ packages/core/primitives/devtools/index.ts
@@ -6,5 +6,9 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-export {ProfilerEvent, type Profiler} from './src/profiler_types';
+export type {ContainerType} from './src/container_type';
+export type {DebugSignalGraphNode} from './src/debug_signal_graph_node';
+export type {Descriptor} from './src/descriptor';
 export {Framework} from './src/framework';
+export {ProfilerEvent, type Profiler} from './src/profiler_types';
+export {PropType} from './src/prop_type';
