+++ packages/core/primitives/devtools/src/container_type.ts
@@ -1,0 +1,9 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.dev/license
+ */
+
+export type ContainerType = 'WritableSignal' | 'ReadonlySignal' | null;
