+++ packages/core/primitives/devtools/src/descriptor.ts
@@ -1,0 +1,19 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.dev/license
+ */
+
+import type {ContainerType} from './container_type';
+import type {PropType} from './prop_type';
+
+export interface Descriptor {
+  expandable: boolean;
+  value?: any;
+  editable: boolean;
+  type: PropType;
+  preview: string;
+  containerType: ContainerType;
+}
