+++ packages/core/primitives/devtools/src/prop_type.ts
@@ -1,0 +1,28 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.dev/license
+ */
+
+export enum PropType {
+  Number,
+  String,
+  Null,
+  Undefined,
+  Symbol,
+  HTMLNode,
+  Boolean,
+  BigInt,
+  Function,
+  Object,
+  Date,
+  Array,
+  Set,
+  Map,
+  Unknown,
+
+  // Special Type when an error occurs during property access
+  Error,
+}
